components relocated
