+++ src/pages/DisplayTasks/DisplayTask.tsx
@@ -1,0 +1,13 @@
+import DisplayTasksComponent from "../../components/Tasks/DisplayTasksComponent";
+
+
+const DisplayTasks = () => {
+
+    return (
+        <div className='main_content'>
+            <DisplayTasksComponent />
+        </div>
+    )
+}
+
+export default DisplayTasks;
