+++ src/pages/ListOfTasks/ListOfTask.tsx
@@ -1,0 +1,12 @@
+import ListOfTasksComponent from "../../components/Tasks/ListOfTasksComponent";
+
+const ListOfTasks = () => {
+
+    return (
+        <div className='main_content'>
+            <ListOfTasksComponent />
+        </div>
+    )
+}
+
+export default ListOfTasks;
